backend/requirements.txt
django
djangorestframework
django-cors-headers
waitress
whitenoise
